rename rules core
