+++ sympy/rules/__init__.py
@@ -27,6 +27,6 @@ import rl
 import traverse
 from rl import rm_id, unpack, flatten, sort, glom, distribute, rebuild
 from util import new
-from strat import (canon, condition, debug, typed, chain, null_safe, do_one,
-        exhaust, minimize)
+from core import condition, debug, chain, null_safe, do_one, exhaust, minimize
+from tools import canon, typed
 import branch
