+++ sympy/rules/tests/test_core.py
@@ -78,3 +78,11 @@ def test_minimize():
 
     rl = minimize(inc, dec, objective=lambda x: -x)
     assert rl(4) == 5
+
+def test_do_one():
+    rl1 = lambda x: 2 if x == 1 else x
+    rl2 = lambda x: 3 if x == 2 else x
+
+    rule = do_one(rl1, rl2)
+    assert rule(1) == 2
+    assert rule(rule(1)) == 3
