+++ sympy/rules/traverse.py
@@ -1,6 +1,6 @@
 """ Strategies to Traverse a Tree """
 from util import new, is_leaf
-from sympy.rules.strat_pure import chain, do_one
+from sympy.rules.core import chain, do_one
 
 def top_down(rule):
     """ Apply a rule down a tree running it on the top nodes first """
